launcher: update to use pups for generating cli flags.
This moves all of the runtime argument generation for docker to use pups
instead of bash. The intention is that launcher should be as simple as
possible as it currently is hard to maintain and add features to. Pups
make sense as we already do processing here and extending it is much
more straightforward.

Launcher instead will lose the smarts for links, volums, env, etc and
will mainly just determine what mode it is running in and behave
accordingly.

In addition, the committed local_discourse/app image, which is an image
baked after the pups commands have been processed, will no longer store
any environment variables, links, or labels as it would have prior. This
has a benefit of not having any secrets stored on disk, they'll only be
provided at run time via the ENV.

It's important that the `docker run` command is ultimately executed
with `bash -c` to avoid issues with quoting of words within the command.

Some changes to launcher itself:
- Move template concatenation to common function, separate run arg gen.
  This was repeated a lot so can belong in it's own function.